eeer

+++ threads.c
@@ -31,10 +31,10 @@ void	*philo_routine(void *pointer) // funkcja przyjmuje argument "pointer", ktor
 
 	philo = (t_philo *)pointer; // pointer jest konwertowany na wskaznik typu "t_philo *", co oznacza, ze pointer jest rzutowany na wskaznik do struktury "t_philo". W ten sposob funkcja uzyskuje dostep do danych konkretnego filozofa.
 	if(philo->id % 2 == 0) // jezeli "id" filozofa jest parzyste...
-		ft_usleep(1); // (tu chyba mozna po prostu napisac usleep(1000) zamiast tworzyc nowa funkcje)... to usypiamy go na 1 milisekunde; robimy to aby unikac ryzyka zakleszczen, czyli nachodzenia na siebie watkow. Dzieki temu unikamy ryzyka, ze dwoch filozofow siegnie po te same widelce w tym samym czasie
-	while(!dead_loop(philo)) // dopoki flozof nie jest martwy (nie zakonczyl dzialania)
+		ft_usleep(1); // (tu chyba mozna po prostu napisac usleep(1000) zamiast tworzyc nowa funkcje)... to usypiamy go na 1 milisekunde; robimy to aby unikac ryzyka zakleszczen, czyli nachodzenia na siebie watkow. Dzieki temu unikamy ryzyka, ze dwoch filozofow siegnie po ten sam widelec w tym samym czasie
+	while(!dead_loop(philo)) // ta petla dziala dopoki filozof nie umrze (zyje=0, martwy=1), czyli dopoki "dead_loop" nie zwroci 1
 	{
-		eat(philo);
+		eat(philo); // filozof probuje podniesc widelce (czyli zablokowac mutexy), zjesc, a nastpenie odlozyc widelce (odblokowac mutexy)
 		dream(philo);
 		think(philo);
 	}
@@ -49,7 +49,7 @@ int	thread_create(t_program *program, pthread_mutex_t *forks)
 	pthread_t	observer; // to jest dodatkowy watek, ktory monitoruje prace filozofow
 	int			x;
 
-	pthread_create((&observer, NULL, &monitor, program->philos) != 0) // pthread_create tworzy watek obserwatora, ktory bedzie wykonywal funkcje "monitor" i przekaze ja jako argument "program->philos". Jesli utworzy go z sukcesem to zwraca 0. W kazdym innym przypadku jest blad.
+	if(pthread_create((&observer, NULL, &monitor, program->philos) != 0)) // pthread_create tworzy watek obserwatora, ktory bedzie wykonywal funkcje "monitor" i przekaze ja jako argument "program->philos". Jesli utworzy go z sukcesem to zwraca 0. W kazdym innym przypadku jest blad.
 		destroy_all("Thread creation error", program, forks); // Wiec jesli funkcja zwraca cos innego niz 1 to jest to blad i "niszcze" wszystko.
 	x = 0;
 	while(x < program->philos[0].num_of_philos) // petla wykonuje sie dopoki x jest mniejsze od ilosci filozofow (ta informacja jest przechowywana w philos[0].num_of_philos, czyli na pierwszym miejscu tablicy philos[0].num_of_philos)
